+++ packages/scenes-app/src/demos/dynamicPage.tsx
@@ -18,6 +18,7 @@ export function getDynamicPageDemo(): SceneAppPage {
     title: 'Dynamic page',
     subTitle: 'Dynamic tabs, and drilldowns. Adds a tab with drilldown after 2 seconds.',
     url: `${demoUrl('dynamic-page')}`,
+    $timeRange: new SceneTimeRange(),
     controls: [new SceneTimePicker({ isOnCanvas: true }), new SceneRefreshPicker({ isOnCanvas: true })],
     tabs: defaultTabs,
   });
@@ -56,7 +57,6 @@ function getSceneAppPage(url: string, name: string) {
             }),
           ],
         }),
-        $timeRange: new SceneTimeRange(),
         $data: getQueryRunnerWithRandomWalkQuery(),
       });
     },
